requirements.txt
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
PyYAML>=6.0
pandas>=2.0.0
scapy>=2.5.0
